bloquear

+++ PMP 2026 - 1/Second_Partial/project/redsocial.cpp
@@ -3,6 +3,7 @@
 #include "redsocial.hpp"
 #include <cstdlib>
 #include "usuario.hpp"
+
 using namespace std;
 //Constructor por Nombre solamente
 RedSocial::RedSocial(string n)
@@ -14,26 +15,33 @@ RedSocial::RedSocial(string n) {
     numeroDePublicaciones = 0;
 }
 */
+
 //Constructor por Nombre y vector de usuarios
 RedSocial::RedSocial(string n, vector<Usuario*> u)
-    : usuarios(u), nombre(n), numeroDeUsuarios(static_cast<int>(u.size())), numeroDePublicaciones(0) {}
+    : usuarios(u), nombre(n), numeroDeUsuarios(u.size()), numeroDePublicaciones(0) {}
+
 //Destructor, donde borra todos los usuarios y publicaciones
 RedSocial::~RedSocial() {
     borrarPublicaciones();
+
     for (size_t i = 0; i < usuarios.size(); i++) {
         delete usuarios[i];
     }
     usuarios.clear();
     numeroDeUsuarios = 0;
 }
+
 //Agregar un puntero Usuario al vector de usuario existente
 void RedSocial::agregarUsuario(Usuario* a) {
     if (a == nullptr) {
         return;
     }
+
     usuarios.push_back(a);
-    numeroDeUsuarios = static_cast<int>(usuarios.size());
+    numeroDeUsuarios = usuarios.size();
 }
+
+
 //Mostrar todos los usuarios de la red social
 void RedSocial::mostrarUsuarios() {
     for (size_t i = 0; i < usuarios.size(); i++) {
@@ -43,6 +51,9 @@ void RedSocial::mostrarUsuarios() {
         cout << "---------" << endl;
     }
 }
+
+
+
 //Mostrar todas las publicaciones de la red social
 void RedSocial::mostrarPublicaciones() {
     for (size_t i = 0; i < usuarios.size(); i++) {
@@ -52,6 +63,7 @@ void RedSocial::mostrarPublicaciones() {
         cout << "---------" << endl;
     }
 }
+
 //obtener un puntero Usuario por su ID
 Usuario* RedSocial::getUsuario(int id) {
     for (size_t i = 0; i < usuarios.size(); i++) {
@@ -59,6 +71,7 @@ Usuario* RedSocial::getUsuario(int id) {
             return usuarios[i];
         }
     }
+
     cout << "No existe un usuario con el id " << id << "." << endl;
     return nullptr;
 }
@@ -69,9 +82,12 @@ Usuario* RedSocial::getUsuario(string nombreBuscado) {
             return usuarios[i];
         }
     }
+
     cout << "No existe un usuario con el nombre " << nombreBuscado << "." << endl;
     return nullptr;
 }
+
+
 //Funcion extra: Borra todas las publicaciones de la red social
 void RedSocial::borrarPublicaciones() {
     for (size_t i = 0; i < usuarios.size(); i++) {
@@ -82,6 +98,8 @@ void RedSocial::borrarPublicaciones() {
             usuarios[i]->publicaciones.clear();
         }
     }
+
     publicaciones.clear();
     numeroDePublicaciones = 0;
 }
+
